refactor: move predict form components to /predict-form
