+++ components/TournamentAdminPanel.tsx
@@ -6,9 +6,9 @@ import { useState } from "react"
 import { api } from "../lib/web/trpc"
 import { useUserId } from "../lib/web/utils"
 import { FixTournamentQuestionSharing } from "./FixTournamentQuestionSharing"
-import { Predict } from "./Predict"
 import { QuestionsMultiselect } from "./QuestionsMultiselect"
 import { ShareTournament } from "./ShareTournament"
+import { Predict } from "./predict-form/Predict"
 import { InfoButton } from "./ui/InfoButton"
 import { Username } from "./ui/Username"
