+++ components/predict-form/Predict.tsx
@@ -18,14 +18,14 @@ import { mergeRefs } from "react-merge-refs"
 import TextareaAutosize from "react-textarea-autosize"
 import SuperJSON from "trpc-transformer"
 import { z } from "zod"
-import { fatebookUrl } from "../lib/_constants"
-import { getDateYYYYMMDD, tomorrowDate } from "../lib/_utils_common"
-import { api } from "../lib/web/trpc"
-import { signInToFatebook, utcDateStrToLocalDate } from "../lib/web/utils"
-import QuestionSuggestions from "./questions/QuestionSuggestions"
-import { QuestionTypeSelect } from "./questions/QuestionTypeSelect"
-import BinaryQuestion from "./questions/question-types/BinaryQuestion"
-import MultiChoiceQuestion from "./questions/question-types/MultiChoiceQuestion"
+import { fatebookUrl } from "../../lib/_constants"
+import { getDateYYYYMMDD, tomorrowDate } from "../../lib/_utils_common"
+import { api } from "../../lib/web/trpc"
+import { signInToFatebook, utcDateStrToLocalDate } from "../../lib/web/utils"
+import QuestionSuggestions from "./QuestionSuggestions"
+import { QuestionTypeSelect } from "./QuestionTypeSelect"
+import BinaryQuestion from "./question-types/BinaryQuestion"
+import MultiChoiceQuestion from "./question-types/MultiChoiceQuestion"
 
 type CreateQuestionMutationOutput = NonNullable<
   ReturnType<typeof api.question.create.useMutation>["data"]
