+++ components/predict-form/PredictModal.tsx
@@ -3,12 +3,12 @@
 import { XCircleIcon } from "@heroicons/react/20/solid"
 import { useSession } from "next-auth/react"
 import React, { useCallback, useEffect, useRef, useState } from "react"
-import { copyToClipboard } from "../lib/web/clipboard"
-import { sendToHost } from "../lib/web/embed"
+import { copyToClipboard } from "../../lib/web/clipboard"
+import { sendToHost } from "../../lib/web/embed"
 import {
   makeClipboardHtmlLink,
   makeRichGoogleDocsLink,
-} from "../lib/web/gdoc_rich_text"
+} from "../../lib/web/gdoc_rich_text"
 import { Predict } from "./Predict"
 
 export default function PredictModal() {
