+++ pages/predict-your-year/[tournament_id].tsx
@@ -7,12 +7,12 @@ import Link from "next/link"
 import { useRouter } from "next/router"
 import { useEffect, useState } from "react"
 import { toast } from "react-hot-toast"
-import { Predict } from "../../components/Predict"
 import { Questions } from "../../components/Questions"
 import { ShareTournament } from "../../components/ShareTournament"
 import { TabbedQuestionSuggestions } from "../../components/TabbedQuestionSuggestions"
 import { TournamentAdminPanel } from "../../components/TournamentAdminPanel"
 import { TournamentLeaderboard } from "../../components/TournamentLeaderboard"
+import { Predict } from "../../components/predict-form/Predict"
 import { generateRandomId } from "../../lib/_utils_common"
 import { api } from "../../lib/web/trpc"
 import { signInToFatebook } from "../../lib/web/utils"
